+++ routes/frontend/home.php
@@ -1,6 +1,7 @@
 <?php
 
-use App\Domains\Auth\Http\Controllers\Frontend\HomeController;
+use App\Http\Controllers\Frontend\HomeController;
+use App\Http\Controllers\Frontend\TermsController;
 use App\Domains\Auth\Http\Controllers\Frontend\User\AccountController;
 use App\Domains\Auth\Http\Controllers\Frontend\User\DashboardController;
 use App\Domains\Auth\Http\Controllers\Frontend\User\ProfileController;
@@ -11,6 +12,7 @@ use App\Domains\Auth\Models\User;
  * All route names are prefixed with 'frontend.'.
  */
 Route::get('/', [HomeController::class, 'index'])->name('index');
+Route::get('terms', [TermsController::class, 'index'])->name('pages.terms');
 
 /*
  * These frontend controllers require the user to be logged in